fix: first part of bugfixes

+++ ewm/src/main/java/ru/practicum/services/CategoryService.java
@@ -6,9 +6,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import ru.practicum.dto.CategoryDto;
 import ru.practicum.dto.news.NewCategoryDto;
+import ru.practicum.exceptions.Conflict;
 import ru.practicum.exceptions.NotFoundException;
 import ru.practicum.mappers.CategoryMapper;
 import ru.practicum.repositories.CategoryRepository;
+import ru.practicum.repositories.EventRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -18,11 +20,12 @@ import java.util.stream.Collectors;
 @Slf4j
 public class CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     public List<CategoryDto> get(Integer from, Integer size) {
         return categoryRepository
                 .findAll(PageRequest.of(from, size)).get()
-                .map(category -> CategoryMapper.toCategoryDto(category))
+                .map(CategoryMapper::toCategoryDto)
                 .collect(Collectors.toList());
     }
 
@@ -35,14 +38,28 @@ public class CategoryService {
     }
 
     public CategoryDto addNewCategory(NewCategoryDto categoryDto) {
+        if (categoryRepository.getCategoryByName(categoryDto.getName()) != null) {
+            log.error("Duplicate category name!");
+            throw new Conflict("Duplicate category name!");
+        }
         return CategoryMapper.toCategoryDto(categoryRepository.save(CategoryMapper.toCategory(categoryDto)));
     }
 
     public void deleteById(Long id) {
+        if (!eventRepository.findByCategoryId(id).isEmpty()) {
+            log.warn("Category being deleted contains events");
+            throw new Conflict("Category being deleted contains events!");
+        }
         categoryRepository.deleteById(id);
     }
 
     public CategoryDto updateById(Long id, CategoryDto dto) {
+        if (dto.getName() != null) {
+            if (categoryRepository.getCategoryByName(dto.getName()) != null) {
+                log.error("Duplicate category name!");
+                throw new Conflict("Duplicate category name!");
+            }
+        }
         return CategoryMapper.toCategoryDto(categoryRepository.save(CategoryMapper.toCategory(id, dto)));
     }
 }
